requirements.txt
annotated-types==0.7.0
clickhouse-driver==0.2.9
dnspython==2.8.0
et_xmlfile==2.0.0
Faker==37.8.0
kafka-python==2.2.15
neo4j==5.28.2
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
